Fixed typo.

+++ include/arctk/data/histogram/average1.hpp
@@ -90,7 +90,7 @@ namespace arc //! arctk namespace
              *  @pre    Size_ must be positive.
              */
             template <typename T>
-            inline Average<T, 1>::Average(const double min_, const double max_, const double size_) noexcept
+            inline Average<T, 1>::Average(const double min_, const double max_, const size_t size_) noexcept
               : Histogram<T, 1>(min_, max_, size_)
               , _counts(size_)
             {
